#5 Parse SGML files
This patch uses our currently implemented text processing utilities to generate CSV files from the original raw SGML data.

We have two files.

One is a CSV containing all data. The data is not normalized and contains everything (except for entries where the text bodies were empty or missing).

The second CSV contains just topic data + the texts that the topics correspond to. This is formatted in a way such that each topic has a column. This is intended to make one-hot vector generation easier.