#11 Track best global performanace.
We now track the hyperparameters and loss and F1 score of the best performing model we've ever gotten into global_best_model_score.json. This will be revision-backed so that we can train on arbitrary hyperparameters and notice when we've gotten a better score when the global_best_model_score.json changes.

This patch also fixes several NaN issues when calculating the F1 score.