#12 Delay bs4 imports
Importing bs4 is difficult to make easily portable since conda doesn't install it easily when using an environment.yml ; thus, we decided to only import it when absolutely necessary, i.e. when we're actually parsing the SGML files. Since we have the preprocessed files on the repo, there's very infrequently the need to preprocess the data. This is a workaround, but it'll unblock us for now. We're considering using setuptools to get bs4 installed.