#13 Reimplement attention layers.
This patch reimplements an attention layer.

It was previously removed bc we thought that 99% accuracy meant we didn't need it.

Turns out that we had a terrible F1 score and we should've checked the true positive count before getting rid of the attention layer.

Either way, we have it back now.