#15 Implement "quit early upon convergence" feature in hyperparameter search.
Our initial experiments have shown that our model trains very slowly.

Thus, we initially made our hyperparameter search train for a large number of epochs.

However, some hyperparameter selections are very bad and we can know after a few epochs (e.g. 5 or so) that the model has converged at some really bad score.

Thus, we shouldn't bother training for 40 epochs if we know after 8 that the model isn't improving.

This patch makes it so that when our hyperparameter search is training on a hyperparmeter set and no validation F1 improvement has been made over NUMBER_OF_RELEVANT_RECENT_EPOCHS (which is currently 5) epochs, then we quit early and move on to the next hyperparamter set.