#16 handle numbers more intelligently via string pre-processing.
This patch makes the data pre-processing space out numeric characters.

This will force each digit to be traeted as a word.

The hope is that we can learn the meaning of the digits and what they mean in sequence.

The idea is that the LSTM going from left to right may hopefully learn what the digits mean.

The right to left LSTM probably won't learn anything.

However, the attention mechanism may account for the digit and it's neighbors.

Using this code snippet, we were able to reduce the number of unknown words from 21375 to 7346 for glove.6B.100d

        from main import * ; from models import *
        spec = 'glove.6B.100d'
        #for spec in ['charngram.100d', 'fasttext.en.300d', 'fasttext.simple.300d', 'glove.42B.300d', 'glove.840B.300d', 'glove.twitter.27B.25d', 'glove.twitter.27B.50d', 'glove.twitter.27B.100d', 'glove.twitter.27B.200d', 'glove.6B.50d', 'glove.6B.100d', 'glove.6B.200d', 'glove.6B.300d']:
        if True:
          classifier = EEAPClassifier(NUMBER_OF_EPOCHS, BATCH_SIZE, TRAIN_PORTION, VALIDATION_PORTION, TESTING_PORTION, 25_000, spec, ENCODING_HIDDEN_SIZE, NUMBER_OF_ENCODING_LAYERS, ATTENTION_INTERMEDIATE_SIZE, NUMBER_OF_ATTENTION_HEADS, DROPOUT_PROBABILITY, OUTPUT_DIR)
          print('\n')
          dpf.spec
          unk_words = sorted(reduce(set.union , ({token for token in example.text if classifier.text_field.vocab.stoi[token] == classifier.unk_idx} for example in classifier.all_data)))
          num_unk_words = len(unk_words)
          dpf.num_unk_words
          p1(random.sample(unk_words, 100))
          dpf.num_unk_words