#16 Add helper method EEAPClassifier.determine_training_unknown_words
While working on #16, we decided that it'd be useful to have a method of determining what words were UNK words from our dataset.

We added EEAPClassifier.determine_training_unknown_words to make this easier.

The torchtext inferface didn't make this easy.

Thus, there's some redundant code we had to use since the torchtext interface didn't support methods with the semantics we wanted. We had to violate thir leaky abstraction to get what we needed.

We were able to determine the number of <unk> words for each set of pre-trained vectors when there was a max vocab size of 10_000.

    charngram.100d : 0
    fasttext.en.300d : 2673
    fasttext.simple.300d : 9032
    glove.42B.300d : 2682
    glove.840B.300d : 4878
    glove.twitter.27B.25d : 6597
    glove.twitter.27B.50d : 6597
    glove.twitter.27B.100d : 6597
    glove.twitter.27B.200d : 6597
    glove.6B.50d : 4137
    glove.6B.100d : 4137
    glove.6B.200d : 4137
    glove.6B.300d : 4137

We used this code as helper code (in addition to some inline changes we made to models.py not included in this commit):

    NUMBER_OF_EPOCHS = 40
    BATCH_SIZE = 1
    MAX_VOCAB_SIZE = 10_000
    TRAIN_PORTION, VALIDATION_PORTION, TESTING_PORTION = (0.50, 0.20, 0.3)

    PRE_TRAINED_EMBEDDING_SPECIFICATION = 'glove.840B.300d'
    ENCODING_HIDDEN_SIZE = 512
    NUMBER_OF_ENCODING_LAYERS = 2
    ATTENTION_INTERMEDIATE_SIZE = 16
    NUMBER_OF_ATTENTION_HEADS = 32
    DROPOUT_PROBABILITY = 0.5

    OUTPUT_DIR = './default_output/'

    from models import *
    from main import *
    for pre_trained_embedding_specification in ['charngram.100d', 'fasttext.en.300d', 'fasttext.simple.300d', 'glove.42B.300d', 'glove.840B.300d', 'glove.twitter.27B.25d', 'glove.twitter.27B.50d', 'glove.twitter.27B.100d', 'glove.twitter.27B.200d', 'glove.6B.50d', 'glove.6B.100d', 'glove.6B.200d', 'glove.6B.300d']:
        print(f'Working on {pre_trained_embedding_specification}')
        classifier = EEAPClassifier(NUMBER_OF_EPOCHS, BATCH_SIZE, TRAIN_PORTION, VALIDATION_PORTION, TESTING_PORTION, MAX_VOCAB_SIZE, pre_trained_embedding_specification, ENCODING_HIDDEN_SIZE, NUMBER_OF_ENCODING_LAYERS, ATTENTION_INTERMEDIATE_SIZE, NUMBER_OF_ATTENTION_HEADS, DROPOUT_PROBABILITY, OUTPUT_DIR)
        print(f'{pre_trained_embedding_specification} : {len(classifier.training_unk_words)}')