#7 Address 2 pieces of technical debt.
There were two software engineering debts accrued early on in order to make progress on more challenging matters.

The first was that we were using some tokenizer from some third party that we were just hoping would sync up nicely with the tokenizer used to generate our training data. That's been fixed. We now use the same tokenizer used to generate our training data.

The second was that we were relying on the string name aliases pointing to word embeddings to contain information about the dimensionality of the vectors. This is brittle. We now dereference the alias and get the dimensionality from there.