#18 Filter out topics with insufficiently many samples.
There are some topics with 1-5 texts associated with them. We cannot predict these topics if the sample count is so small. They're so small that they're not even guaranteed to show up in our sample set.

Thus, we remove these topics.

Keep in mind, the text is still there. The text might have other topics attached, so we should still learn from the texts from that perspective.