#20 Speed Up F1 Threshold Optimization.
Optimizing the threshold right now takes a really long time.

It's because it uses the batch size of the training iterator.

We want to keep that low since our model takes a long time to train.

However, we don't a low batch size for optimizing the F1 threshold since we're not doing any SGD. We just want the computation to be fast. Ideally, the batch size for the F1 threshold optimization would be the size of the training set.

We set it to 32. This will make determining the optimal F1 threshold (which is really just a bunch of sums) faster.