#17 Implement Data Augmentation via UNK Replacements
This commit implements a data augmentation method that takes inputs sequences and randomly replaces some tokens with the UNK token.

This hasn't shown to significantly increase performance, so we'll leave the data augmentation probability at zero.

This commit also contains an implementation of the soft recall loss, which also hasn't shown to increase performance.

This patch also fixes a printing issue with tqdm_with_messsage that was causing there to be excessive white space being shown at the end of everyline.

This patch also fixes a bug in our logging of the globally best performing model. Dynamically generated functions weren't having their names printed properly (e.g. this can happen with decorated functions), so we addressed that.