#4 Finish first draft "Data Description & Problem Overview"
This commit completes the first section of write up that goes over the problem definition, a brief and broad overview of the data, and an explanation of our pre-processing pipeline.

I added in some details about what our expected interface will be.

I showed some examples of what the pre-processed data looks like.

I also found that our pre-processed data was out of date, so this patch updates that as well.

There are many more changes.