#4 Abstract out Classifier abstract class to separate module
Since we're presenting this project as done, it's very important to do all the last minor bits of abstraction that we can to make the code nice and pretty.

This moves out the Classifier abstract class to abstract_classifier.py.

It's only used by models.py.

models.py now only contains torch.nn models and concrete implementations of the abstract Classifier.

models.py now is more inline with what people might think of as the concrete implementation of a model, i.e. torch.nn.Module instances.