#8 Add third example to document, fix argparse spacing, update data
This commit does several things.

This makes the CLI look nicer by adding the correct amount of spacing so that the help options are more readable.

This adds a third example to describe what the desired product is in our documentation.

This fixes a bunch of CSS issues with scroll bars showing up where they are not needed.