#2 Complete README with pointers to full documentations
This completes the first pass at writing the README.

There's not much to write. We simply want to state our findings and point to the full documentation.

We showed the results.

We stated some libraries we used.

We told the reader to contact us if they have issues.

Hopefully, this will be a sufficient README.