#8 Fix footercard display to be more mobile friendly
This commit makes the footer cards have the image and the text not overlap.

This commit additionally makes the display of the output labels in the docs for the desired interface have optional scroll bars.

Both of these are only necessary for mobile devices where the width is significantly smaller than the height.